Fixed type conversion bug.

+++ api/handlers.go
@@ -66,7 +66,7 @@ func (s *server) HandleCategorySummary() http.HandlerFunc {
 
 		// loop through categories and get 3 latest entries on that category
 		for _, cat := range categories {
-			entries, err := s.db.GetCategoryEntries(context.Background(), int(cat["id"].(int64)), 0, 3)
+			entries, err := s.db.GetCategoryEntries(context.Background(), cat["id"].(int), 0, 3)
 			if err != nil {
 				continue
 			}
